@giovp 2021-07-14
Great catch @marcovarrone ! thank you for the PR!
As mentioned by michal, only minor thing in pre-commit checks. I'd suggest you to `pre-commit install` and then commit again the file, the checks will run and fix the error. Let us know for anything else.