@marcovarrone 2021-07-14
Hi @michalk8 and @giovp, thank you for the feedback.
I had some problems with running pre-commit (I fixed it and I will submit another PR for it soon).

I hope everything is fine now!